Extract client SDK package
