+++ packages/cli/package.json
@@ -24,6 +24,7 @@
   },
   "dependencies": {
     "@clack/prompts": "^1.0.0",
+    "@getpaseo/client": "0.1.83",
     "@getpaseo/server": "0.1.83",
     "chalk": "^5.3.0",
     "commander": "^12.0.0",
