+++ packages/cli/src/commands/agent/logs.ts
@@ -2,7 +2,8 @@ import { Command } from "commander";
 import { connectToDaemon, getDaemonHost } from "../../utils/client.js";
 import type { CommandOptions } from "../../output/index.js";
 import { fetchProjectedTimelineItems } from "../../utils/timeline.js";
-import type { DaemonClient, AgentStreamMessage, AgentTimelineItem } from "@getpaseo/server";
+import type { DaemonClient } from "@getpaseo/client/internal/daemon-client";
+import type { AgentStreamMessage, AgentTimelineItem } from "@getpaseo/server";
 import { curateAgentActivity } from "@getpaseo/server";
 
 export function addLogsOptions(cmd: Command): Command {
