+++ packages/cli/src/commands/agent/stop.ts
@@ -1,5 +1,5 @@
 import { Command } from "commander";
-import type { DaemonClient } from "@getpaseo/server";
+import type { DaemonClient } from "@getpaseo/client/internal/daemon-client";
 import { connectToDaemon, getDaemonHost } from "../../utils/client.js";
 import { isSameOrDescendantPath } from "../../utils/paths.js";
 import type {
