+++ packages/cli/tests/e2e/relay-host.test.ts
@@ -6,10 +6,10 @@ import path from "node:path";
 import { fileURLToPath } from "node:url";
 import {
   buildRelayWebSocketUrl,
-  DaemonClient,
   generateLocalPairingOffer,
   parseConnectionOfferFromUrl,
 } from "@getpaseo/server";
+import { DaemonClient } from "@getpaseo/client/internal/daemon-client";
 import { WebSocket } from "ws";
 import { getAvailablePort } from "../helpers/network.ts";
 import { createE2ETestContext } from "../helpers/test-daemon.ts";
