+++ packages/server/package.json
@@ -63,6 +63,7 @@
   "dependencies": {
     "@agentclientprotocol/sdk": "^0.17.1",
     "@anthropic-ai/claude-agent-sdk": "^0.2.133",
+    "@getpaseo/client": "0.1.83",
     "@getpaseo/highlight": "0.1.83",
     "@getpaseo/relay": "0.1.83",
     "@isaacs/ttlcache": "^2.1.4",
