+++ packages/server/src/server/exports.ts
@@ -19,13 +19,6 @@ export {
 } from "../shared/connection-offer.js";
 export { buildRelayWebSocketUrl } from "../shared/daemon-endpoints.js";
 export {
-  DaemonClient,
-  type DaemonClientConfig,
-  type ConnectionState,
-  type DaemonEvent,
-} from "../client/daemon-client.js";
-export type { WebSocketFactory, WebSocketLike } from "../client/daemon-client-transport-types.js";
-export {
   buildDaemonWebSocketUrl,
   deriveLabelFromEndpoint,
   normalizeHostPort,
